+++ CHANGES.rst
@@ -13,6 +13,7 @@ Bugfixes
 - Point to correct day in "edit session timetable" link (:issue:`3419`)
 - Fix error when exporting abstracts with review questions to JSON
 - Point the timetable to correct day in the session details
+- Fix massive performance issue on the material package page in big events
 
 Version 2.1.2
 -------------